@@ -4,7 +4,7 @@
   "description": "",
   "main": "main.js",
   "scripts": {
-    "lint": "eslint src",
+    "lint": "eslint src --ext .ts",
     "test": "npx tsc && mocha",
     "coverage": "nyc mocha",
     "pretest": "npm run lint",
@@ -26,7 +26,9 @@
     "@types/lodash": "^3.10.4",
     "@types/node": "^17.0.12",
     "@types/screeps": "^3.2.4",
-    "eslint": "^6.8.0",
+    "@typescript-eslint/eslint-plugin": "^5.10.2",
+    "@typescript-eslint/parser": "^5.10.2",
+    "eslint": "^7.1.0",
     "eslint-plugin-screeps": "^2.1.0",
     "lodash": "^3.10.1",
     "mocha": "^7.1.1",
