@@ -7,7 +7,7 @@ import rBr = require("./breaker")
 import settings = require("../config/settings")
 import cN = require("../lib/creepNames")
 
-var rPM = {
+const rPM = {
     name: cN.POWER_MINER_NAME,
     type: "powerMiner",
     boosts: [RESOURCE_CATALYZED_GHODIUM_ALKALIDE, RESOURCE_CATALYZED_UTRIUM_ACID],
